+++ src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
 import Footer from './components/Footer';
-import Navbar from './components/Navbar';
+import Navbar from './components/navbar/Navbar';
 
 function App() {
   return (
